vet_assist_ai/requirements.txt
fastapi
uvicorn
mysql-connector-python
pandas
langchain
langchain-openai
python-dotenv
pydantic